+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed node and port highlighting when creating a draft connection - #362
 - Fixed duplicate node creation on some desktop systems - #350
 - Fixed icons not being licensed under Apache 2.0.
+- Fixed memory leak in the `GraphUserVariablesDialog`.
 
 ### Added
 - Selected nodes can now be "cut" using the corresponding shortcut (usually `Ctrl+X`). Cut-nodes are "greyed-out", similar to how cutting works for files and folders. Once the selection is pasted, the cut-objects are deleted. - #283
 - The style of comments can be customized using a dedicated dialog that can be opened using the context menu of comments. This dialog allows changing the background color and text color, frame visibility, and text alignment. - #326
+- The Graph User Variables Dialog can now be opened on any "User Constants" (`GraphUserVariablesInputNode`) Node - #360
 
 ### Changed